+++ pmd-apex/src/main/java/net/sourceforge/pmd/lang/apex/metrics/AbstractApexMetric.java
@@ -4,12 +4,16 @@
 
 package net.sourceforge.pmd.lang.apex.metrics;
 
+import net.sourceforge.pmd.lang.ast.Node;
+import net.sourceforge.pmd.lang.metrics.AbstractMetric;
+
 /**
  * Base class for all Apex metrics.
  *
  * @author Clément Fournier
  */
-public class AbstractApexMetric {
+public abstract class AbstractApexMetric<N extends Node> extends AbstractMetric<N> {
+
 
     protected ApexSignatureMatcher getSignatureMatcher() {
         return ApexMetrics.getFacade().getProjectMirror();